updated calendar page after merging with connected frontend
