+++ app/(CustomerPages)/calendar/layout.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+
+export default function CalendarLayout({
+    children,
+}:  {
+    children: React.ReactNode
+}
+)   {
+    return (
+        <section>
+          <main>{children}</main>
+        </section>
+      );
+}
