--- app/calendar/page.tsx
@@ -1,15 +1,0 @@
-'use client'
-
-import React from 'react'
-import MyCalendar from './Components/Calendar'
-
-export default function Calendar() {
-  return (
-    <div>
-      <MyCalendar />
-    </div>
-  )
-};
-
-
-
